fix: mobile bindings missing Orb options
