+++ cmd/agent-mobile/pkg/wrappers/command/testdata/contexts/credentials-examples_v1.jsonld
