+++ cmd/agent-mobile/pkg/wrappers/command/testdata/contexts/odrl.jsonld
